+++ docs/changelog.rst
@@ -44,6 +44,8 @@ Changed
 
 - :meth:`xugrid.Ugrid2d.from_structured` now takes ``x`` and ``y`` arguments instead
   of ``x_bounds`` and ``y_bounds`` arguments.
+- :func:`xugrid.merge_partitions` allows merging partitions with different grids
+  per partition.
 
 [0.8.1] 2024-01-19
 ------------------