+++ docs/changelog.rst
@@ -6,8 +6,8 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog`_, and this project adheres to
 `Semantic Versioning`_.
 
-Unreleased
-----------
+[0.9.0] 2024-02-15
+------------------
 
 Fixed
 ~~~~~